Add 'scripts' folder that holds all the scripts that will be copied in the target root filesystem.
Add 'scripts/lis3_read.sh' script to read a set of resisters from
the lis3lv02d inertial sensor. A sample set of registers is stored
in 'lis3_registers'.
On the target system, 'lis3_registers' can either be edited, or can
be used as a template to create a new file with a custom set of
registers.

  usage: lis3_read <registers_filename>