'i2c-stn8815.c': add flushing of the i2c FIFO buffers at startup.
As a step of the controller initialization, the 'stn8815_i2c_hwinit'
function flushes the Rx and Tx FIFOs of the I2C controller. As reported
in the controller datasheet, the flushing phase takes several system
clock cycles to complete. Thus, the driver polls the flush status
in a small busy-waiting loop (up to 10ms).
In case of timeout reports an error.