'i2c-stn8815': Fix 'Unbalanced pm_runtime_enable!'
This message is issued by the kernel if the driver is unloaded
(i.e. using 'rmmod') and then loaded back (insmod).
The cause is that pm_runtime is enabled on probing, but it's never
disabled (neither if probing fails, nor if the driver is removed).

Fix by disabling pm_runtime on driver unloading and on probing
failure (only if it was already enabled).